frontend: Store stateless clusters info in indexDB
This removes initial storage of kubeconfig data from session storage to
indexDB. This change was made for maximum data retention since indexDB
has much more data storage then session storage. Also we are not storing
headlamp-userId in localstorage so that when a loggedin user open
another tab they are able to see their stateless clusters.
