+++ frontend/package.json
@@ -48,6 +48,7 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-plugin-unused-imports": "^1.1.5",
+    "fake-indexeddb": "^5.0.1",
     "github-markdown-css": "^5.1.0",
     "https-browserify": "^1.0.0",
     "humanize-duration": "^3.27.2",
