+++ frontend/src/lib/k8s/apiProxy.ts
@@ -347,9 +347,9 @@ export async function clusterRequest(
   let fullPath = path;
   if (cluster) {
     const token = getToken(cluster);
-    const matchingKubeconfig = findKubeconfigByClusterName(cluster);
-    if (matchingKubeconfig !== null) {
-      opts.headers['KUBECONFIG'] = matchingKubeconfig;
+    const kubeconfig = await findKubeconfigByClusterName(cluster);
+    if (kubeconfig !== null) {
+      opts.headers['KUBECONFIG'] = kubeconfig;
       opts.headers['X-HEADLAMP-USER-ID'] = userID;
     }
 
@@ -1369,13 +1369,15 @@ function connectStreamWithParams(
   if (cluster) {
     fullPath = combinePath(`/${CLUSTERS_PREFIX}/${cluster}`, path);
     // Include the userID as a query parameter if it's a stateless cluster
-    const matchingKubeconfig = findKubeconfigByClusterName(cluster);
-    if (matchingKubeconfig !== null) {
-      const queryParams = `X-HEADLAMP-USER-ID=${userID}`;
-      url = combinePath(BASE_WS_URL, fullPath) + (fullPath.includes('?') ? '&' : '?') + queryParams;
-    } else {
-      url = combinePath(BASE_WS_URL, fullPath);
-    }
+    findKubeconfigByClusterName(cluster).then(kubeconfig => {
+      if (kubeconfig !== null) {
+        const queryParams = `X-HEADLAMP-USER-ID=${userID}`;
+        url =
+          combinePath(BASE_WS_URL, fullPath) + (fullPath.includes('?') ? '&' : '?') + queryParams;
+      } else {
+        url = combinePath(BASE_WS_URL, fullPath);
+      }
+    });
   }
 
   let socket: WebSocket | null = null;
@@ -1571,7 +1573,7 @@ export async function setCluster(clusterReq: ClusterRequest) {
   const kubeconfig = clusterReq.kubeconfig;
 
   if (kubeconfig) {
-    storeStatelessClusterKubeconfig(kubeconfig);
+    await storeStatelessClusterKubeconfig(kubeconfig);
     return;
   }
